typecast to long for THCCeilDiv in THCReduce.cuh

+++ lib/THC/THCReduce.cuh
@@ -321,7 +321,7 @@ bool THC_reduceDim(THCState* state,
           block.y *= 2;
           ydim /= 2;
         }
-        THC_getGridFromTiles(THCCeilDiv(outElements, (long)block.x), grid);
+        THC_getGridFromTiles(THCCeilDiv((long)outElements, (long)block.x), (long)grid);
 
     }
   }
